SearchImg Function moved to Icons folder
SearchImg function moved into icons folder.  Both CharIcon and RankIcon utilize the logic from this function.
